reworked SwipeToDelete and documented it
