+++ app/src/main/java/com/example/clicker/presentation/stream/StreamView.kt
@@ -332,243 +332,3 @@ fun TextChat(
 }
 
 
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun SwipeToDeleteChatMessages(
-    twitchUser: TwitchUserData,
-    bottomModalState: ModalBottomSheetState,
-    updateClickedUser: (String, String, Boolean, Boolean) -> Unit,
-    deleteMessage: (String) -> Unit,
-
-
-) {
-    ChatCard(
-        twitchUser = twitchUser,
-        bottomModalState = bottomModalState,
-        updateClickedUser = { username, userId, banned, isMod ->
-            updateClickedUser(
-                username,
-                userId,
-                banned,
-                isMod
-            )
-        },
-        deleteMessage = { messageId -> deleteMessage(messageId) },
-
-    )
-}
-
-
-
-@OptIn(ExperimentalMaterialApi::class, ExperimentalFoundationApi::class)
-@Composable
-fun ChatCard(
-    twitchUser: TwitchUserData,
-    bottomModalState: ModalBottomSheetState,
-    updateClickedUser: (String, String, Boolean, Boolean) -> Unit,
-    deleteMessage: (String) -> Unit,
-) {
-    val subBadge = "https://static-cdn.jtvnw.net/badges/v1/5d9f2208-5dd8-11e7-8513-2ff4adfae661/1"
-    val modBadge = "https://static-cdn.jtvnw.net/badges/v1/3267646d-33f0-4b17-b3df-f923a41db1d0/1"
-    val coroutineScope = rememberCoroutineScope()
-
-    var color by remember { mutableStateOf(Color(parseColor(twitchUser.color))) }
-    if(color == Color.Black){
-        color = androidx.compose.material3.MaterialTheme.colorScheme.primary
-    }
-
-    val state = rememberSwipeableActionsState()
-
-    var offset = state.offset.value
-
-    val swipeThreshold = 130.dp
-    val swipeThresholdPx = LocalDensity.current.run { swipeThreshold.toPx() }
-
-    val thresholdCrossed = abs(offset) > swipeThresholdPx
-
-    var backgroundColor by remember { mutableStateOf(Color.Black) }
-    var fontSize = 17.sp
-
-    if (thresholdCrossed) {
-        backgroundColor = Color.Red
-
-    } else {
-        backgroundColor = androidx.compose.material3.MaterialTheme.colorScheme.primary
-    }
-
-//    if(showOneClickAction){
-//        offset = 0f
-//    }
-    // makes it so mods can not be swiped on
-    val modDragState = DraggableState { delta ->
-    }
-
-    var dragState = state.draggableState
-    if (twitchUser.mod == "1") {
-        dragState = modDragState
-    }
-    if (twitchUser.deleted) {
-        dragState = modDragState
-        backgroundColor = Color.Red
-        fontSize = 14.sp
-    }
-
-    val cardWidth = Resources.getSystem().displayMetrics.widthPixels.dp // width of what will be moving
-    val scope = rememberCoroutineScope()
-    val primary = androidx.compose.material3.MaterialTheme.colorScheme.primary
-
-    Log.d("deleteChatMessageException", "WTwitchUser.userId---> ${twitchUser.userId}")
-
-    DetectSwipeBox(
-        backgroundColor =backgroundColor,
-        dragState =dragState,
-        state =state,
-        thresholdCrossed =thresholdCrossed,
-        deleteMessage ={messageId ->deleteMessage(messageId)},
-        twitchUserId = twitchUser.id,
-        cardThatMoves={
-            ClickableCard(
-                twitchUser =twitchUser,
-                color = color,
-                bottomModalState = bottomModalState,
-                offset = offset,
-                fontSize = fontSize,
-                updateClickedUser = {  username, userId,isBanned,isMod ->
-                    updateClickedUser(
-                        username,
-                        userId,
-                        isBanned,
-                        isMod
-                    )
-                }
-            )
-        }
-    )
-}
-
-@Composable
-fun DetectSwipeBox(
-    backgroundColor:Color,
-    dragState: DraggableState,
-    state: SwipeableActionsState,
-    thresholdCrossed: Boolean,
-    deleteMessage: (String) -> Unit,
-    twitchUserId:String?,
-    cardThatMoves:@Composable () -> Unit,
-
-){
-    val scope = rememberCoroutineScope()
-    Box(
-        Modifier
-            .fillMaxWidth()
-            .padding(vertical = 10.dp, horizontal = 10.dp)
-            .background(backgroundColor)
-            .draggable(
-                orientation = Orientation.Horizontal,
-                enabled = true,
-                state = dragState,
-                onDragStopped = {
-                    scope.launch {
-
-                        if (thresholdCrossed) {
-                            state.resetOffset()
-                            deleteMessage(twitchUserId ?: "")
-                        } else {
-                            state.resetOffset()
-                        }
-                    }
-                },
-                onDragStarted = {
-
-                }
-
-            )
-
-    ) {
-        cardThatMoves()
-    }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun ClickableCard(
-    twitchUser: TwitchUserData,
-    color: Color,
-    offset: Float,
-    bottomModalState: ModalBottomSheetState,
-    fontSize:TextUnit,
-    updateClickedUser: (String, String, Boolean, Boolean) -> Unit,
-
-){
-    val coroutineScope = rememberCoroutineScope()
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .absoluteOffset { IntOffset(x = offset.roundToInt(), y = 0) }
-            .clickable {
-                updateClickedUser(
-                    twitchUser.displayName.toString(),
-                    twitchUser.userId.toString(),
-                    twitchUser.banned,
-                    twitchUser.mod != "1"
-                )
-                coroutineScope.launch {
-                    bottomModalState.show()
-                }
-            },
-        backgroundColor = androidx.compose.material3.MaterialTheme.colorScheme.primary,
-        border = BorderStroke(2.dp, androidx.compose.material3.MaterialTheme.colorScheme.secondary)
-
-    ) {
-        Column() {
-            CheckIfUserDeleted(twitchUser = twitchUser)
-            CheckIfUserIsBanned(twitchUser = twitchUser)
-            TextWithChatBadges(
-                twitchUser = twitchUser,
-                color = color,
-                fontSize = fontSize
-            )
-        }
-    }
-}
-
-@Composable
-fun CheckIfUserDeleted(twitchUser: TwitchUserData){
-    if (twitchUser.deleted) {
-        Text(
-            stringResource(R.string.moderator_deleted_comment),
-            fontSize = 20.sp,
-            modifier = Modifier.padding(start = 5.dp),
-            color = MaterialTheme.colors.onPrimary
-        )
-    }
-}
-@Composable
-fun CheckIfUserIsBanned(twitchUser: TwitchUserData){
-    if (twitchUser.banned) {
-        val duration = if (twitchUser.bannedDuration != null) "Banned for ${twitchUser.bannedDuration} seconds" else "Banned permanently"
-        Text(duration, fontSize = 20.sp, modifier = Modifier.padding(start = 5.dp))
-    }
-}
-@Composable
-fun TextWithChatBadges(
-    twitchUser: TwitchUserData,
-    color:Color,
-    fontSize: TextUnit
-
-){
-    Row(
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-
-        ChatBadges(
-            username = "${twitchUser.displayName} :",
-            message = " ${twitchUser.userType}",
-            isMod = twitchUser.mod == "1",
-            isSub = twitchUser.subscriber == true,
-            color = color,
-            textSize = fontSize
-        )
-
-    } // end of the row
-}
