+++ app/src/main/java/com/example/clicker/presentation/stream/views/MainChat.kt
@@ -470,89 +470,7 @@ object MainChat{
 
     }
 }
-@Composable
-fun ChatBadges(
-    username: String,
-    message: String,
-    isMod: Boolean,
-    isSub: Boolean,
-    color: Color,
-    textSize: TextUnit
-) {
-    //for not these values can stay here hard coded. Until I implement more Icon
-    val modBadge = "https://static-cdn.jtvnw.net/badges/v1/3267646d-33f0-4b17-b3df-f923a41db1d0/1"
-    val subBadge = "https://static-cdn.jtvnw.net/badges/v1/5d9f2208-5dd8-11e7-8513-2ff4adfae661/1"
-    val modId = "modIcon"
-    val subId = "subIcon"
-    val text = buildAnnotatedString {
-        // Append a placeholder string "[icon]" and attach an annotation "inlineContent" on it.
-        if (isMod) {
-            appendInlineContent(modId, "[icon]")
-        }
-        if (isSub) {
-            appendInlineContent(subId, "[subicon]")
-        }
-        withStyle(style = SpanStyle(color = color, fontSize = textSize)) {
-            append(username)
-        }
-        withStyle(style = SpanStyle(color = MaterialTheme.colorScheme.onPrimary)) {
-            append(message)
-        }
-    }
 
-    val inlineContent = mapOf(
-        Pair(
-
-            modId,
-            InlineTextContent(
-
-                Placeholder(
-                    width = 20.sp,
-                    height = 20.sp,
-                    placeholderVerticalAlign = PlaceholderVerticalAlign.Center
-                )
-            ) {
-                AsyncImage(
-                    model = modBadge,
-                    contentDescription = stringResource(R.string.moderator_badge_icon_description),
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(2.dp)
-                )
-            }
-        ),
-        Pair(
-
-            subId,
-            InlineTextContent(
-
-                Placeholder(
-                    width = 20.sp,
-                    height = 20.sp,
-                    placeholderVerticalAlign = PlaceholderVerticalAlign.Center
-                )
-            ) {
-                AsyncImage(
-                    model = subBadge,
-                    contentDescription = stringResource(R.string.sub_badge_icon_description),
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(2.dp)
-                )
-            }
-        )
-    )
-
-    Text(
-        text = text,
-        inlineContent = inlineContent,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(5.dp),
-        color = color,
-        fontSize = textSize
-    )
-}
